# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AZM CRM API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LocalizedText": {
        "type": "object",
        "description": "Bilingual text (constitution I). Every refusal message uses this shape.",
        "properties": {
          "ar": {
            "type": "string"
          },
          "en": {
            "type": "string"
          }
        },
        "required": [
          "ar",
          "en"
        ]
      },
      "Refusal": {
        "type": "object",
        "properties": {
          "message": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Field names that failed validation, where applicable."
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/customer": {
      "get": {
        "summary": "Ranked-match customer search (spec 001 FR-002; the single-identity-key question was closed as malformed — see decisions-pending §0 decision 7)",
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Minimum 3 characters (E-13) — matched against phone, email, national ID, account reference and display name, in that rank order"
          }
        ],
        "responses": {
          "200": {
            "description": "Results scope-filtered; below 3 characters returns an empty, non-error result (E-13)"
          }
        }
      },
      "post": {
        "summary": "Create a customer from a display name and one contact point (spec 001 FR-001, AS-01)",
        "tags": [
          "Customer"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "displayName",
                  "contactPoints"
                ],
                "properties": {
                  "displayName": {
                    "type": "string",
                    "description": "User-authored, single language — not a localised value"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "person",
                      "organisation"
                    ],
                    "default": "person"
                  },
                  "nationalId": {
                    "type": "string",
                    "nullable": true
                  },
                  "accountRef": {
                    "type": "string",
                    "nullable": true,
                    "description": "The ERP join key. Optional — the CRM is the system of record (decision 6)."
                  },
                  "preferredLanguage": {
                    "type": "string",
                    "enum": [
                      "ar",
                      "en"
                    ]
                  },
                  "organisationId": {
                    "type": "string",
                    "nullable": true
                  },
                  "contactPoints": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "channelType",
                        "value"
                      ],
                      "properties": {
                        "channelType": {
                          "type": "string",
                          "enum": [
                            "phone",
                            "email",
                            "whatsapp"
                          ]
                        },
                        "value": {
                          "type": "string"
                        },
                        "isPrimary": {
                          "type": "boolean"
                        }
                      }
                    }
                  },
                  "confirmCollision": {
                    "type": "boolean",
                    "description": "Repeat the request with this set to true to proceed past a 409 collision (decision 16 — FR-004 unique-by-default, overridable)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Missing or invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "404": {
            "description": "organisationId does not resolve to a visible organisation"
          },
          "409": {
            "description": "A matching customer already exists (FR-010, AS-04). Not blocked — resend with confirmCollision true.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "matches": {
                      "type": "array",
                      "description": "In-scope matches, named"
                    },
                    "outOfScopeMatches": {
                      "type": "integer",
                      "description": "Out-of-scope collisions, counted but never named (spec 010 §8)"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customer/{id}": {
      "get": {
        "summary": "Get one customer (spec 001 FR-003, AS-03 — out of scope is 404, never 403)",
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found — identical whether out of scope or genuinely absent"
          }
        }
      },
      "patch": {
        "summary": "Edit customer fields (spec 001 FR-020 — one audit entry per changed field)",
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  },
                  "nationalId": {
                    "type": "string"
                  },
                  "accountRef": {
                    "type": "string"
                  },
                  "preferredLanguage": {
                    "type": "string",
                    "enum": [
                      "ar",
                      "en"
                    ]
                  },
                  "preferredChannel": {
                    "type": "string",
                    "enum": [
                      "phone",
                      "email",
                      "whatsapp"
                    ]
                  },
                  "sensitiveFlag": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "No changes supplied"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
...
}